Assorted cleanup in indexing.cpp
Originally committed to SVN as r385.

+++ src/core/indexing.h
@@ -123,7 +123,7 @@ protected:
 	TAudioNameCallback ANC;
 	void *ANCPrivate;
 	const char *SourceFile;
-	AlignedBuffer<int16_t> DecodingBuffer;
+	AlignedBuffer<uint8_t> DecodingBuffer;
 
 	int64_t Filesize;
 	uint8_t Digest[20];
